# parser: use non-capturing groups