Should avoid many small copies of the input we never use.

Version regex capture groups were fixed in #40344.

As an example, the `spack.parser.ALL_TOKENS` regex has 69 (!) capture groups. Running it on

```
%gcc@1,2:4,5:
```

gives 14 matches... that's rather excessive.

Combining this and #40344 parsing gets about 8-10% faster, benchmarked on:

```
$ wc specs.txt  # one spec per line
  1000   3708 119013 specs.txt
```

```python
# before
In [1]: import spack.parser

In [2]: %timeit spack.parser.SpecParser(open("specs.txt").read()).all_specs()
74.6 ms ± 230 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```

```python
# after
In [1]: import spack.parser

In [2]: %timeit spack.parser.SpecParser(open("specs.txt").read()).all_specs()
68.5 ms ± 273 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```